+++ include/shared/Overlay.h
@@ -6,6 +6,6 @@
 
 void InitializeOverlay(HWND hwnd);
 void CleanupOverlay();
-void DrawOverlay(HWND hwnd, double angle, bool showCrosshair);
+void DrawOverlay(HWND hwnd, double angle, bool showCrosshair, bool overlayVisible = true);
 
 #endif // OVERLAY_H